+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-Changes since v1.0.2 will be documented here.
+### Added
+
+- Added transparent `.gz` and `.zst` reads for persisted file APIs when the
+  corresponding compression feature is enabled.
 
 ## [v1.0.2] - 2026-09-19